pyproject.toml
[tool.poetry]
name = "devops_entrega_1"
version = "0.1.0"
description = ""
authors = ["Juan Sebastian Rodriguez Moreno <js.rodriguezm12345@uniandes.edu.co>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
flask = "^3.1.2"
flask-marshmallow = "^1.3.0"
flask-sqlalchemy = "^3.1.1"
flask-restful = "^0.3.10"
flask-jwt-extended = "^4.7.1"
werkzeug = "^3.1.3"
marshmallow-sqlalchemy = "^1.4.2"
psycopg2-binary = "^2.9.11"
gunicorn = "^23.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
